Move get_table_columns_metric func to great_expectations.expectations.metrics.util
