+++ tests/execution_engine/test_sparkdf_execution_engine.py
@@ -11,6 +11,7 @@ from great_expectations.core.batch_spec import PathBatchSpec, RuntimeDataBatchSp
 from great_expectations.core.metric_domain_types import MetricDomainTypes
 from great_expectations.core.metric_function_types import MetricPartialFunctionTypes
 from great_expectations.execution_engine import SparkDFExecutionEngine
+from great_expectations.expectations.metrics.util import get_table_columns_metric
 from great_expectations.expectations.row_conditions import (
     RowCondition,
     RowConditionParserType,
@@ -18,7 +19,6 @@ from great_expectations.expectations.row_conditions import (
 from great_expectations.self_check.util import build_spark_engine
 from great_expectations.validator.computed_metric import MetricValue
 from great_expectations.validator.metric_configuration import MetricConfiguration
-from tests.expectations.test_util import get_table_columns_metric
 from tests.test_utils import create_files_in_directory
 from great_expectations.optional_imports import F, sparktypes, pyspark_sql_Row
